requirements.txt
astroid==2.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.12
django-import-export==2.5.0
et-xmlfile==1.0.1
isort==5.8.0
lazy-object-proxy==1.6.0
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
pkg-resources==0.0.0
pylint==2.7.4
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
